requirements.txt
# Core Web / API
fastapi
uvicorn[standard]
python-multipart
pydantic

# Frontend / UI
streamlit>=1.28.0
plotly
matplotlib
seaborn
markdown

# Computer Vision / Media
opencv-python
Pillow
mediapipe
moviepy==1.0.3
imageio-ffmpeg

# AI / ML Core
torch
torchvision
numpy
pandas
scikit-learn
lightgbm
joblib
accelerate>=0.26.0
tqdm

# NLP / LLM
transformers
datasets
sentence-transformers
langchain
langchain-openai
google-generativeai
openai
openai-whisper
pdfplumber
python-dotenv